git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@10779 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/dialheap.h
@@ -40,7 +40,7 @@ class CDialHeap {
 		inline tPathNode* Route (uint i = 0) { return m_route.Buffer (i); }
 
 	private:
-		ushort CDialHeap::Scan (short* buffer, int nStart, int nLength);
+		int CDialHeap::Scan (short* buffer, int nStart, int nLength);
 };
